+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - Fix SSH authentication log always showing "no authentication".
 
+### Changed
+
+- Minor improvements to admin interface.
+
 ## 0.2.0 (2024-12-20)
 
 ### Added